requirements.txt
 fastapi==0.114.2
 uvicorn[standard]==0.30.6
 pydantic-settings==2.6.0
 pydantic==2.9.2
 psycopg2-binary==2.9.9
 pgvector==0.2.5
 SQLAlchemy==2.0.36
 httpx==0.27.2
 structlog==24.1.0
 opentelemetry-api==1.27.0
 opentelemetry-sdk==1.27.0
 opentelemetry-exporter-otlp==1.27.0
 opentelemetry-instrumentation-fastapi==0.48b0
 prometheus-fastapi-instrumentator==6.1.0
 python-dotenv==1.0.1
 tiktoken==0.8.0
 openai==1.51.0
- sentence-transformers==3.2.1
 numpy==2.1.2
